+++ test/widgets/journal/entry_details/delete_icon_widget_test.dart
@@ -34,6 +34,7 @@ void main() {
         (_) => EntryState.dirty(
           entryId: testTextEntry.meta.id,
           entry: testTextEntry,
+          showMap: false,
         ),
       );