+++ test/widgets/journal/entry_details/entry_datetime_widget_test.dart
@@ -31,6 +31,7 @@ void main() {
         (_) => EntryState.dirty(
           entryId: testTextEntry.meta.id,
           entry: testTextEntry,
+          showMap: false,
         ),
       );
     });